# perf: don't build schema in to_dict