## Problem
Did some profiling and extending the pydantic schema the way we were doing it was taking a long time. This was the largest cause of slowness for the dashboard api.

## Changes
Just do this manipulation as a dict. We were making the schema just to dump it back to a dict anyways, so simple enough.

Benchmarked it locally and this speeds tile generation by a factor of 3.5x.

I do have another change that switches the DRF serializes out with just creating raw dictionaries which profiles another 2x speed up. I will put that in a different PR and we can consider if that's worth it.

Pasting in Claude's nice table:

● 🎉 INCREDIBLE RESULTS! Complete Performance Comparison:

  | Test                       | Time per Call | Total Time (20 calls) | Speedup vs Baseline | Key Bottlenecks                                      |
  |----------------------------|---------------|-----------------------|---------------------|------------------------------------------------------|
  | BASELINE                   | 1.154s        | 23.1s                 | 1.0x                | Pydantic model construction (13.8s), DRF serializers |
  | to_dict Only               | 0.327s        | 6.5s                  | 🚀 3.5x             | DRF serializers (3.2s), remaining field construction |
  | to_dict + fast_serializers | 0.171s        | 3.4s                  | 🚀 6.7x             | Database queries, minimal serialization overhead     |

  🏆 FINAL PERFORMANCE ACHIEVEMENTS:

  Overall Improvement: 6.7x Faster!

  - From 1.154s → 0.171s per dashboard call
  - From 23.1s → 3.4s for 20 calls
  - 85% reduction in total time

## How did you test this code?
Ran unit tests and tested locally.